+++ source/ggSwarm/ggSwarm/tasks/direct/ggswarm_marl/__init__.py
@@ -26,3 +26,13 @@ gym.register(
         "skrl_mappo_cfg_entry_point": f"{agents.__name__}:skrl_mappo_cfg.yaml",
     },
 )
+
+gym.register(
+    id="GGS-Hover-v0",
+    entry_point=f"{__name__}.drone_hover_env:GGSwarmHoverEnv",
+    disable_env_checker=True,
+    kwargs={
+        "env_cfg_entry_point": f"{__name__}.drone_hover_env_cfg:GGSwarmHoverEnvCfg",
+        "skrl_mappo_cfg_entry_point": f"{agents.__name__}:skrl_mappo_hover_cfg.yaml",
+    },
+)